backend-tests/common: factor out common functions
actually move common stuff to common.py.

changelog: none
